Move VLM test utilities to sglang.test.vlm_utils
